@@ -1,18 +1,16 @@
-apply plugin: 'java'
-apply plugin: 'war'
-apply plugin: 'eclipse-wtp'
+plugins {
+	id 'java'
+	id 'war'
+	id 'eclipse-wtp'
+	id 'org.iplass.dev' version '1.0.0'
+}
 
 ext {
-	javaVersion = JavaVersion.VERSION_1_8
+	javaVersion = JavaVersion.VERSION_21
 }
 
 repositories {
 	mavenCentral()
-
-	maven {
-		// for jasper reports
-		url 'https://jaspersoft.jfrog.io/jaspersoft/third-party-ce-artifacts/'
-	}
 
 	if (project.hasProperty('iPLAssMavenRepoUrl')) {
 		maven {
