@@ -39,7 +39,7 @@ import samples.ec01.utils.Consts;
 @ActionMapping(
 		name = "samples/ec01/top", 
 		displayName = "デモトップ画面表示アクション", 
-		privilaged = true, 
+		privileged = true, 
 		result = @Result(
 				type = Type.TEMPLATE, 
 				value = "samples/ec01/top"))
