@@ -35,7 +35,7 @@ import samples.ec01.utils.Consts;
 @ActionMapping(
 		name = "samples/ec01/shipping/inputShippingInfoNoMember", 
 		displayName = "非会員配送情報入力画面表示アクション", 
-		privilaged = true, 
+		privileged = true, 
 		result = {
 				@Result(
 						status = Constants.CMD_EXEC_ERROR, 
